+++ src/viewers/dxf-viewer/worker.ts
@@ -1,0 +1,2 @@
+import { DxfViewer } from 'dxf-viewer';
+DxfViewer.SetupWorker();